archive_v1/requirements_legacy_full.txt
alembic==1.17.2
altgraph==0.17.5
annotated-types==0.7.0
anyio==4.12.0
arrow==1.4.0
attrs==25.4.0
bandit==1.9.2
binaryornot==0.4.4
black==25.12.0
blinker==1.9.0
cachelib==0.13.0
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cookiecutter==2.6.0
coverage==7.13.0
croniter==6.0.0
cryptography==46.0.3
Deprecated==1.3.1
distro==1.9.0
execnet==2.1.2
filetype==1.2.0
flake8==7.3.0
flasgger==0.9.7.1
Flask==3.1.2
Flask-Caching==2.3.1
flask-cors==6.0.2
Flask-Limiter==4.1.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
flask-talisman==1.1.0
flet==0.28.3
flet-cli==0.28.3
flet-desktop==0.28.3
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.41.1
google-auth-httplib2==0.2.1
google-auth-oauthlib==1.2.3
google-genai==1.55.0
google-generativeai==0.8.5
googleapis-common-protos==1.72.0
greenlet==3.3.0
grpcio==1.76.0
grpcio-status==1.71.2
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
langchain-core==0.3.80
langchain-google-genai==2.0.10
langsmith==0.4.53
limits==5.6.0
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mccabe==0.7.0
mdurl==0.1.2
mistune==3.2.0
mypy_extensions==1.1.0
oauthlib==3.3.1
ordered-set==4.1.0
orjson==3.11.4
packaging==25.0
pathspec==0.12.1
pefile==2024.8.26
pillow==12.0.0
platformdirs==4.5.1
pluggy==1.6.0
prometheus_client==0.23.1
proto-plus==1.26.1
protobuf==5.29.5
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.14.0
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pyflakes==3.4.0
pygame==2.6.1
Pygments==2.19.2
pyinstaller==6.17.0
pyinstaller-hooks-contrib==2025.10
pyOpenSSL==25.3.0
pyparsing==3.2.5
pypng==0.20220715.0
pytest==9.0.2
pytest-cov==7.0.0
pytest-mock==3.15.1
pytest-order==1.3.0
pytest-xdist==3.8.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.2.1
python-json-logger==4.0.0
python-slugify==8.0.4
pytokens==0.3.0
pytz==2025.2
pywin32==311
pywin32-ctypes==0.2.3
python-telegram-bot==21.9
pyTelegramBotAPI==4.14.0
openai>=1.0.0
pystray>=0.19.5
PyYAML==6.0.3
qrcode==7.4.2
redis==7.1.0
referencing==0.37.0
repath==0.9.0
reportlab==4.4.7
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.2.0
rpds-py==0.30.0
rq==2.6.1
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.45
stevedore==5.6.0
structlog==25.5.0
tenacity==9.1.2
text-unidecode==1.3
toml==0.10.2
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uuid==1.30
uuid_utils==0.12.0
watchdog==4.0.2
websockets==15.0.1
Werkzeug==3.1.4
WMI==1.5.1
wrapt==2.0.1
zstandard==0.25.0
edge-tts>=7.0.0
pydub>=0.25.0
opencv-python>=4.10.0
numpy>=2.3.0,<3.0.0
schedule>=1.2.0
psutil>=5.9.0
ddgs>=0.1.0
pyautogui>=0.9.54
pygetwindow>=0.0.9
pvporcupine>=2.2.1
pyaudio>=0.2.13
faster-whisper>=0.10.0
mediapipe>=0.10.0
python-osc>=1.8.0